source var Git pull, also fix egrep bug
Now the Git pull checks if there's a dir by the name $name-git, if there
isn't then it clones, if there is then it goes into the dir and pulls
from remote, deleting any chnages that may have been made in tree.

The egrep bug was that we weren't piping anything into egrep, so it
stay there, waiting for eternity.  Now we echo the source into it.