agnoadd and agnomk got some niceties
agnomk:
Now checks if file exists before trying to download it, if it does exist
then it skips the download and informs the user.

agnoadd:
Checks if provided package file exists, exits if it doesn't.

It also checks if the file is installed (i.e. can be found in
$AGNO_META), if it is installed then it first uninstalls it and then
installs it again.

bootstrap.sh:
ln -s -> ln -sf.  I just figured out that the commands I had under
usr/bin/ weren't symlinks, but rather copies of the original files (for
whatever reason), so the last commits I wasn't actually testing.  I did
this in order to make sure the symlinking actually does something.